Unify Pen & HyperPen tools
Now there's just a single tool that has a flag set to determine
whether new paths are bezier or hyperbezier.
